fix an issue where db seeder doesnt have access to test factories
