+++ lib/oli/utils/db_seeder.ex
@@ -22,7 +22,6 @@ defmodule Oli.Seeder do
   alias Oli.Qa.Reviews
   alias Oli.Activities
   alias Oli.Delivery.Gating
-  alias Oli.Factory
 
   def base_project_with_resource(author) do
     {:ok, family} =
@@ -1375,26 +1374,5 @@ defmodule Oli.Seeder do
       })
 
     gating_condition
-  end
-
-  def create_author_account(), do: create_user_account(SystemRole.role_id().author)
-  def create_admin_account(), do: create_user_account(SystemRole.role_id().admin)
-
-  defp create_user_account(role_id) do
-    password =
-      for _ <- 1..20, into: "", do: <<Enum.random('0123456789abcdefghijklmnopqrstuvwxyz_$#@!')>>
-
-    {:ok, user} =
-      Author.changeset(Factory.author_factory(), %{
-        password: password,
-        password_confirmation: password,
-        system_role_id: role_id
-      })
-      |> Repo.insert()
-
-    %{
-      user: user,
-      password: password
-    }
   end
 end
